chore<BE>: 변수 변경
Former-commit-id: 5dd67ef42d6f3ebbe7a37cdd2dc45b940f4c9880

+++ BE/src/main/java/com/ssafy/IBG/api/RecommendApiController.java
@@ -367,7 +367,7 @@ public class RecommendApiController {
     @GetMapping("/game/score/all")
     public Result setScoreRecommend() throws JsonProcessingException {
         try{
-            int MAX_USER = userService.getUserSize();
+            Long MAX_USER = userService.getUserSize();
             for(int userNo = 331993; userNo < MAX_USER; userNo++){
                 restapiService.requestGETAPI("/user3", userNo);
             }
